refactor: Move off of deprecated 'trace' path
